requirements.txt
appnope==0.1.0
asn1crypto==0.24.0
backports.shutil-get-terminal-size==1.0.0
bcrypt==3.1.4
Cerberus==1.1
cffi==1.11.5
configparser==3.5.0
cryptography==1.7.2
decorator==4.2.1
enum34==1.1.6
falcon==1.4.1
gunicorn==19.7.1
idna==2.6
ipaddress==1.0.19
ipdb==0.11
ipython==5.5.0
ipython-genutils==0.2.0
itsdangerous==0.24
jwt==0.5.2
mysqlclient==1.3.12
numpy==1.14.2
opencv-python==3.4.0.12
pathlib2==2.3.0
pexpect==4.4.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pyasn1==0.4.2
pycparser==2.18
Pygments==2.2.0
python-engineio==2.0.4
python-mimeparse==1.6.0
python-socketio==1.9.0
redis==2.10.6
scandir==1.7
shortuuid==0.5.0
simplegeneric==0.8.1
six==1.11.0
SQLAlchemy==1.2.5
traitlets==4.3.2
typing==3.5.3.0
wcwidth==0.1.7
